Move Configuration class to storage module
