+++ clcache.py
@@ -500,31 +500,6 @@ class PersistentJSONDict(object):
         return type(self) is type(other) and self.__dict__ == other.__dict__
 
 
-class Configuration(object):
-    _defaultValues = {"MaximumCacheSize": 1073741824} # 1 GiB
-
-    def __init__(self, configurationFile):
-        self._configurationFile = configurationFile
-        self._cfg = None
-
-    def __enter__(self):
-        self._cfg = PersistentJSONDict(self._configurationFile)
-        for setting, defaultValue in self._defaultValues.items():
-            if setting not in self._cfg:
-                self._cfg[setting] = defaultValue
-        return self
-
-    def __exit__(self, typ, value, traceback):
-        # Does not write to disc when unchanged
-        self._cfg.save()
-
-    def maximumCacheSize(self):
-        return self._cfg["MaximumCacheSize"]
-
-    def setMaximumCacheSize(self, size):
-        self._cfg["MaximumCacheSize"] = size
-
-
 class Statistics(object):
     CALLS_WITH_INVALID_ARGUMENT = "CallsWithInvalidArgument"
     CALLS_WITHOUT_SOURCE_FILE = "CallsWithoutSourceFile"
